[NEW] Refactor fixtures
